Chess_Tournament_Project_SQL_Database
Project Background:
Your organization has sub-branches in 49 countries and its HQ is in Dallas. The club always accepts new players and new clubs. At the beginning of every May, you are starting a new tournament and all clubs must have a participant to attend and rankings will be posted on your web site.

Goal: create a relational database to meet the requirements of this tournament. (MS SQL)